[core] separate FileTree, move into xci/core, add PathNode tests
