+++ tools/find_file/ff.cpp
@@ -8,6 +8,8 @@
 /// A find-like tool using Hyperscan for regex matching.
 
 #include <xci/core/ArgParser.h>
+#include <xci/core/FileTree.h>
+#include <xci/core/container/FlatSet.h>
 #include <xci/core/sys.h>
 #include <xci/core/string.h>
 #include <xci/core/log.h>
@@ -17,223 +19,14 @@
 #include <fmt/core.h>
 #include <hs/hs.h>
 
-#include <memory>
-#include <functional>
-#include <thread>
-#include <mutex>
-#include <condition_variable>
-#include <cassert>
 #include <cstring>
 #include <utility>
 #include <string_view>
 
 #include <sys/types.h>
-#include <sys/stat.h>
-#include <dirent.h>
-#include <fcntl.h>
 
 using namespace xci::core;
 using namespace xci::core::argparser;
-
-
-class FileTree {
-public:
-    struct PathNode {
-        explicit PathNode(std::string_view component) : component(component) {}  // NOLINT
-        PathNode(std::string_view component, const std::shared_ptr<PathNode>& parent) : component(component), parent(parent) {}  // NOLINT
-
-        std::string to_string() const {
-            return dir_to_string() + component;
-        }
-
-        std::string dir_to_string() const {
-            if (!parent)
-                return {};
-            return parent->to_string() + '/';
-        }
-
-        std::string component;
-        int fd = -1;
-        std::shared_ptr<PathNode> parent;
-    };
-
-    enum Type {
-        File,
-        Directory,
-        OpenError,
-        OpenDirError,
-        ReadDirError,
-    };
-
-    /// for Directories, return true to descent, false to skip
-    using Callback = std::function<bool(const PathNode&, Type)>;
-
-    /// \param max_threads      Number of threads FileTree can spawn.
-    explicit FileTree(unsigned max_threads, unsigned queue_size, Callback&& cb) : m_max_threads(max_threads), m_cb(std::move(cb)) {
-        assert(m_cb);
-        m_workers.reserve(max_threads);
-        m_queue.reserve(queue_size);
-        assert(m_queue.capacity() == queue_size);
-    }
-
-    ~FileTree() {
-        // join threads
-        for (auto& t : m_workers)
-            t.join();
-    }
-
-    void walk(const std::string& pathname) {
-        // create PathNode also for parent, so the reporting is consistent
-        // (component in each reported PathNode is always cleaned basename)
-        auto pathname_clean = rstrip(pathname, '/');
-        auto components = rsplit(pathname_clean, '/', 1);
-        std::shared_ptr<PathNode> path;
-        if (components.size() == 2) {
-            auto parent = std::make_shared<PathNode>(components[0]);
-            path = std::make_shared<PathNode>(components[1], parent);
-        } else {
-            path = std::make_shared<PathNode>(pathname_clean);
-        }
-        if (pathname.empty()) {
-            int fd = open(".", O_DIRECTORY | O_NOFOLLOW | O_NOCTTY, O_RDONLY);
-            if (fd == -1) {
-                m_cb(*path, OpenError);
-                return;
-            }
-            path->fd = fd;
-        } else {
-            if (!open_and_report(pathname.c_str(), *path))
-                return;
-        }
-        enqueue(std::move(path));
-    }
-
-    void worker() {
-        TRACE("[{}] worker start", get_thread_id());
-        std::unique_lock lock(m_mutex);
-        while(!m_queue.empty() || m_busy != 0) {
-            while (m_queue.empty()) {
-                m_cv.wait(lock);
-                if (m_busy == 0 && m_queue.empty()) {
-                    lock.unlock();
-                    m_cv.notify_all();
-                    TRACE("[{}] worker finish", get_thread_id());
-                    return;
-                }
-            }
-
-            auto path = std::move(m_queue.back());
-            m_queue.pop_back();
-            TRACE("[{}] worker read start ({} busy, {} queue)", get_thread_id(), m_busy, m_queue.size());
-            ++m_busy;
-            lock.unlock();
-            read(path);
-            lock.lock();
-            --m_busy;
-            TRACE("[{}] worker read finish ({} busy, {} queue)", get_thread_id(), m_busy, m_queue.size());
-        }
-        lock.unlock();
-        m_cv.notify_all();
-        TRACE("[{}] worker finish", get_thread_id());
-    }
-
-private:
-    void start_worker() {
-        m_workers.emplace_back(std::thread{[this]{
-            worker();
-        }});
-    }
-
-    void enqueue(std::shared_ptr<PathNode>&& path) {
-        std::unique_lock lock(m_mutex);
-        if (m_queue.size() < m_queue.capacity()) {
-            m_queue.emplace_back(std::move(path));
-            lock.unlock();
-            m_cv.notify_one();
-        } else {
-            if (m_workers.size() < m_max_threads)
-                start_worker();
-            // process the item in this thread
-            // (better than blocking and doing nothing)
-            TRACE("[{}] enqueue read start ({} busy, {} queue)", get_thread_id(), m_busy, m_queue.size());
-            ++m_busy;
-            lock.unlock();
-            read(path);
-            lock.lock();
-            --m_busy;
-            TRACE("[{}] enqueue read finish ({} busy, {} queue)", get_thread_id(), m_busy, m_queue.size());
-        }
-    }
-
-    void read(const std::shared_ptr<PathNode>& path) {
-        DIR* dirp = fdopendir(path->fd);
-        if (dirp == nullptr) {
-            m_cb(*path, OpenDirError);
-            close(path->fd);
-            return;
-        }
-
-        for (;;) {
-            errno = 0;
-            auto* dir_entry = readdir(dirp);
-            if (dir_entry == nullptr) {
-                // end or error
-                if (errno != 0) {
-                    m_cb(*path, ReadDirError);
-                }
-                break;
-            }
-            if (strcmp(dir_entry->d_name, ".") == 0 || strcmp(dir_entry->d_name, "..") == 0)
-                continue;
-
-            auto entry_path = std::make_shared<PathNode>(dir_entry->d_name, path);
-
-            if ((dir_entry->d_type & DT_DIR) == DT_DIR || dir_entry->d_type == DT_UNKNOWN) {
-                // readdir says it's a dir or it doesn't know
-                if (open_and_report(dir_entry->d_name, *entry_path, path->fd))
-                    enqueue(std::move(entry_path));
-                continue;
-            }
-            m_cb(*entry_path, File);
-        }
-        closedir(dirp);
-    }
-
-    /// \param pathname     path to open, may be relative
-    /// \param node         PathNode associated with the path, for reporting
-    /// \param at_fd        if path is relative, this can be open parent FD or AT_FDCWD
-    /// \return     true if opened as a directory and callback returned true (i.e. descend)
-    bool open_and_report(const char* pathname, PathNode& node, int at_fd = AT_FDCWD) {
-        // try to open as a directory, if it fails with ENOTDIR, it is a file
-        int fd = openat(at_fd, pathname, O_DIRECTORY | O_NOFOLLOW | O_NOCTTY, O_RDONLY);
-        if (fd == -1) {
-            if (errno == ENOTDIR) {
-                // it's a file - report it
-                m_cb(node, File);
-                return false;
-            }
-            if (!m_cb(node, Directory))
-                return false;
-            m_cb(node, OpenError);
-            return false;
-        }
-        node.fd = fd;
-
-        if (!m_cb(node, Directory)) {
-            close(fd);
-            return false;
-        }
-        return true;
-    }
-
-    const unsigned m_max_threads;
-    Callback m_cb;
-    std::vector<std::shared_ptr<PathNode>> m_queue;
-    std::vector<std::thread> m_workers;
-    int m_busy = 0;  // number of threads in `read`
-    std::mutex m_mutex;  // for both queue and workers
-    std::condition_variable m_cv;
-};
 
 
 int main(int argc, const char* argv[])
@@ -242,6 +35,8 @@ int main(int argc, const char* argv[])
     bool ignore_case = false;
     bool show_hidden = false;
     bool show_dirs = false;
+    bool search_in_special_dirs = false;
+    bool single_device = false;
     bool show_version = false;
     int jobs = 8;
     std::vector<const char*> files;
@@ -254,9 +49,12 @@ int main(int argc, const char* argv[])
             Option("-F, --fixed", "Match literal string instead of (default) regex", fixed),
 #endif
             Option("-i, --ignore-case", "Enable case insensitive matching", ignore_case),
-            Option("-H, --show-hidden", "Don't skip hidden files", show_hidden),
-            Option("-D, --show-dirs", "Don't skip directory entries", show_dirs),
-            Option("-a, --all", "Don't skip any files, same as -H -D", [&]{ show_hidden = true; show_dirs = true; }),
+            Option("-H, --search-hidden", "Don't skip hidden files", show_hidden),
+            Option("-D, --search-dirnames", "Don't skip directory entries", show_dirs),
+            Option("-S, --search-in-special-dirs", "Allow descending into special directories like /dev", search_in_special_dirs),
+            Option("-X, --single-device", "Don't descend into directories with different device number. "
+                                          "This has similar effect to -S, but costs additional stat() call per directory.", single_device),
+            Option("-a, --all", "Don't skip any files, same as -HD", [&]{ show_hidden = true; show_dirs = true; }),
             Option("-c, --color", "Force color output", [&]{ term.set_is_tty(TermCtl::IsTty::Always); }),
             Option("-j, --jobs JOBS", "Number of worker threads", jobs).env("JOBS"),
             Option("-V, --version", "Show version", show_version),
@@ -322,8 +120,10 @@ int main(int argc, const char* argv[])
         .highlight = term.bold().yellow().seq(),
     };
 
+    FlatSet<dev_t> dev_ids;
+
     FileTree ft(jobs-1, jobs,
-                [show_hidden, show_dirs, pattern, &re_db, &theme]
+                [show_hidden, show_dirs, single_device, pattern, &re_db, &theme, &dev_ids]
                 (const FileTree::PathNode& path, FileTree::Type t)
     {
         if (!show_hidden && path.component[0] == '.')
@@ -331,7 +131,16 @@ int main(int argc, const char* argv[])
         switch (t) {
             case FileTree::Directory:
                 if (!show_dirs)
-                    return true;
+                    return true;  // descent
+                if (single_device) {
+                    struct stat st;
+                    if (!path.stat(st)) {
+                        fmt::print(stderr,"ff: stat({}): {}\n", path.to_string(), errno_str());
+                        return true;
+                    }
+                    if (!dev_ids.contains(st.st_dev))
+                        return false;  // skip (different device ID)
+                }
                 FALLTHROUGH;
             case FileTree::File:
                 if (pattern) {
@@ -364,10 +173,10 @@ int main(int argc, const char* argv[])
                     std::string out;
                     if (t == FileTree::Directory) {
                         out += theme.dir;
-                        out += path.dir_to_string();
+                        out += path.dirname();
                     } else {
                         out += theme.file_dir;
-                        out += path.dir_to_string();
+                        out += path.dirname();
                         out += theme.file_name;
                     }
                     out += std::string_view(name, so);
@@ -387,7 +196,7 @@ int main(int argc, const char* argv[])
                         out += path.to_string();
                     } else {
                         out += theme.file_dir;
-                        out += path.dir_to_string();
+                        out += path.dirname();
                         out += theme.file_name;
                         out += path.component;
                     }
@@ -407,10 +216,14 @@ int main(int argc, const char* argv[])
         }
         UNREACHABLE;
     });
-    if (files.empty())
+    if (files.empty()) {
+        // TODO: add to dev_ids
         ft.walk("");
-    for (const auto& f : files)
+    }
+    for (const auto& f : files) {
+        // TODO: add to dev_ids
         ft.walk(f);
+    }
 
     ft.worker();
